+++ src/projects/projects.module.ts
@@ -3,9 +3,11 @@ import { ProjectsService } from './projects.service';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { Project } from './entities/project.entity';
 import { UserModule } from '../user/user.module';
+import { ProjectsController } from './projects.controller';
 
 @Module({
   imports: [TypeOrmModule.forFeature([Project]), UserModule],
   providers: [ProjectsService],
+  controllers: [ProjectsController],
 })
 export class ProjectsModule {}
