+++ src/time/time.controller.ts
@@ -1,5 +1,4 @@
 import { Controller, Get } from '@nestjs/common';
 
 @Controller('api/time')
-export class TimeController {
-}
+export class TimeController {}
